+++ packages/transaction-controller/src/TransactionController.ts
@@ -1535,6 +1535,7 @@ export class TransactionController extends BaseController<
       chainId: this.#getChainId(networkClientId),
       ethQuery,
       isSimulationEnabled: this.#isSimulationEnabled(),
+      messenger: this.messagingSystem,
       txParams: transaction,
     });
 
@@ -1562,6 +1563,7 @@ export class TransactionController extends BaseController<
       chainId: this.#getChainId(networkClientId),
       ethQuery,
       isSimulationEnabled: this.#isSimulationEnabled(),
+      messenger: this.messagingSystem,
       txParams: transaction,
     });
 
@@ -4143,6 +4145,7 @@ export class TransactionController extends BaseController<
       ethQuery,
       isCustomNetwork,
       isSimulationEnabled: this.#isSimulationEnabled(),
+      messenger: this.messagingSystem,
       txMeta: transactionMeta,
     });
   }